chore: restructure code - dashboard views - clickhouse
