+++ futurex_openedx_extensions/dashboard/views/clickhouse.py
@@ -1,0 +1,180 @@
+"""Views for the dashboard app - clickhouse feature"""
+# pylint: disable=duplicate-code
+
+from __future__ import annotations
+
+from typing import Any, Dict
+from urllib.parse import parse_qs, urlencode, urlsplit, urlunsplit
+
+from django.core.exceptions import ValidationError
+from django.core.paginator import EmptyPage
+from django.http import JsonResponse
+from edx_api_doc_tools import exclude_schema_for
+from rest_framework import status as http_status
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from futurex_openedx_extensions.helpers import clickhouse_operations as ch
+from futurex_openedx_extensions.helpers.constants import (
+    CLICKHOUSE_FX_BUILTIN_CA_USERS_OF_TENANTS,
+    CLICKHOUSE_FX_BUILTIN_ORG_IN_TENANTS,
+    FX_VIEW_DEFAULT_AUTH_CLASSES,
+)
+from futurex_openedx_extensions.helpers.converters import error_details_to_dictionary
+from futurex_openedx_extensions.helpers.models import ClickhouseQuery
+from futurex_openedx_extensions.helpers.pagination import DefaultPagination
+from futurex_openedx_extensions.helpers.permissions import FXHasTenantCourseAccess
+from futurex_openedx_extensions.helpers.roles import FXViewRoleInfoMixin, get_usernames_with_access_roles
+
+default_auth_classes = FX_VIEW_DEFAULT_AUTH_CLASSES.copy()
+
+
+@exclude_schema_for('get')
+class ClickhouseQueryView(FXViewRoleInfoMixin, APIView):
+    """View to get the Clickhouse query"""
+    authentication_classes = default_auth_classes
+    permission_classes = [FXHasTenantCourseAccess]
+    fx_view_name = 'clickhouse_query_fetcher'
+    fx_default_read_only_roles = ['staff', 'instructor', 'data_researcher', 'org_course_creator_group']
+    fx_view_description = 'api/fx/query/v1/<scope>/<slug>: Get result of the related clickhouse query'
+
+    @staticmethod
+    def get_page_url_with_page(url: str, new_page_no: int | None) -> str | None:
+        """
+        Get the URL with the new page number
+
+        :param url: The URL
+        :type url: str
+        :param new_page_no: The new page number
+        :type new_page_no: int | None
+        :return: The URL with the new page number
+        :rtype: str | None
+        """
+        if new_page_no is None:
+            return None
+
+        url_parts = urlsplit(url)
+        query_params = parse_qs(url_parts.query)
+
+        page_size = query_params.get(DefaultPagination.page_size_query_param, None)
+        if page_size:
+            del query_params[DefaultPagination.page_size_query_param]
+
+        if 'page' in query_params:
+            del query_params['page']
+
+        if page_size:
+            query_params[DefaultPagination.page_size_query_param] = page_size
+        query_params['page'] = [str(new_page_no)]
+
+        new_query_string = urlencode(query_params, doseq=True)
+
+        new_url_parts = (url_parts.scheme, url_parts.netloc, url_parts.path, new_query_string, url_parts.fragment)
+        new_full_url = urlunsplit(new_url_parts)
+        return new_full_url
+
+    @staticmethod
+    def pop_out_page_params(params: Dict[str, str], paginated: bool) -> tuple[int | None, int]:
+        """
+        Pop out the page and page size parameters, and return them as integers in the result. Always return the page
+        as None if not paginated
+
+        :param params: The parameters
+        :type params: Dict[str, str]
+        :param paginated: Whether the query is paginated
+        :type paginated: bool
+        :return: The page and page size parameters
+        :rtype: tuple[int | None, int]
+        """
+        page_str: str | None = params.pop('page', None)
+        page_size_str: str = params.pop(
+            DefaultPagination.page_size_query_param, ''
+        ) or str(DefaultPagination.page_size)
+
+        if not paginated:
+            page = None
+        else:
+            page = int(page_str) if page_str is not None else page_str
+            page = 1 if page is None else page
+
+        return page, int(page_size_str)
+
+    def get(self, request: Any, scope: str, slug: str) -> JsonResponse | Response:
+        """
+        GET /api/fx/query/v1/<scope>/<slug>/
+
+        :param request: The request object
+        :type request: Request
+        :param scope: The scope of the query (course, tenant, user)
+        :type scope: str
+        :param slug: The slug of the query
+        :type slug: str
+        """
+        clickhouse_query = ClickhouseQuery.get_query_record(scope, 'v1', slug)
+        if not clickhouse_query:
+            return Response(
+                error_details_to_dictionary(reason=f'Query not found {scope}.v1.{slug}'),
+                status=http_status.HTTP_404_NOT_FOUND
+            )
+
+        if not clickhouse_query.enabled:
+            return Response(
+                error_details_to_dictionary(reason=f'Query is disabled {scope}.v1.{slug}'),
+                status=http_status.HTTP_400_BAD_REQUEST
+            )
+
+        params = request.query_params.dict()
+        self.get_page_url_with_page(request.build_absolute_uri(), 9)
+
+        page, page_size = self.pop_out_page_params(params, clickhouse_query.paginated)
+
+        orgs = request.fx_permission_info['view_allowed_any_access_orgs'].copy()
+        params[CLICKHOUSE_FX_BUILTIN_ORG_IN_TENANTS] = orgs
+        if CLICKHOUSE_FX_BUILTIN_CA_USERS_OF_TENANTS in clickhouse_query.query:
+            params[CLICKHOUSE_FX_BUILTIN_CA_USERS_OF_TENANTS] = get_usernames_with_access_roles(orgs)
+
+        error_response = None
+        try:
+            clickhouse_query.fix_param_types(params)
+
+            with ch.get_client() as clickhouse_client:
+                records_count, next_page, result = ch.execute_query(
+                    clickhouse_client,
+                    query=clickhouse_query.query,
+                    parameters=params,
+                    page=page,
+                    page_size=page_size,
+                )
+
+        except EmptyPage as exc:
+            error_response = Response(
+                error_details_to_dictionary(reason=str(exc)), status=http_status.HTTP_404_NOT_FOUND
+            )
+        except (ch.ClickhouseClientNotConfiguredError, ch.ClickhouseClientConnectionError) as exc:
+            error_response = Response(
+                error_details_to_dictionary(reason=str(exc)), status=http_status.HTTP_503_SERVICE_UNAVAILABLE
+            )
+        except (ch.ClickhouseBaseError, ValueError) as exc:
+            error_response = Response(
+                error_details_to_dictionary(reason=str(exc)), status=http_status.HTTP_400_BAD_REQUEST
+            )
+        except ValidationError as exc:
+            error_response = Response(
+                error_details_to_dictionary(reason=exc.message), status=http_status.HTTP_400_BAD_REQUEST
+            )
+
+        if error_response:
+            return error_response
+
+        if clickhouse_query.paginated:
+            return JsonResponse({
+                'count': records_count,
+                'next': self.get_page_url_with_page(request.build_absolute_uri(), next_page),
+                'previous': self.get_page_url_with_page(
+                    request.build_absolute_uri(),
+                    None if page == 1 else page - 1 if page else None,
+                ),
+                'results': ch.result_to_json(result),
+            })
+
+        return JsonResponse(ch.result_to_json(result), safe=False)
